+++ backend/eventsbooking/src/main/java/com/application/eventsbooking/models/EventStatus.java
@@ -8,7 +8,7 @@ public class EventStatus {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long statusId;
+    private int statusId;
 
     @ManyToOne
     @JoinColumn(name = "event_id", nullable = false)
